Implemented knight movement and attack of barriers

+++ src/figures/Dwarf.java
@@ -10,7 +10,7 @@ public class Dwarf extends Figure{
      * @param row row position
      * @param col col position
      */
-    public Dwarf(double row, double col,Color color) {
+    public Dwarf(int row, int col,Color color) {
         super(row, col,color);
     }
 
@@ -18,9 +18,9 @@ public class Dwarf extends Figure{
         Graphics2D lineDrawer = (Graphics2D) g;
 
         int widthOfTile = 100;
-        int tileX = (int) (this.col * widthOfTile);
+        int tileX = this.col * widthOfTile;
         int heightOfTile = 100;
-        int tileY = (int) (this.row * heightOfTile);
+        int tileY = this.row * heightOfTile;
 
         g.setColor(color);
         g.fillRect(tileX,tileY, widthOfTile, heightOfTile);
